+++ piccolo/apps/playground/commands/run.py
@@ -5,7 +5,6 @@ for interacting with the data using Piccolo.
 
 import datetime
 import sys
-import typing as t
 import uuid
 from decimal import Decimal
 from enum import Enum
@@ -14,6 +13,7 @@ from piccolo.columns import (
     JSON,
     UUID,
     Boolean,
+    Date,
     ForeignKey,
     Integer,
     Interval,
@@ -104,7 +104,7 @@ class Ticket(Table):
         return Readable(
             template="%s - %s",
             columns=[
-                t.cast(t.Type[Venue], cls.concert.venue).name,
+                cls.concert._.venue._.name,
                 cls.ticket_type,
             ],
         )
@@ -136,7 +136,31 @@ class RecordingStudio(Table):
         )
 
 
-TABLES = (Manager, Band, Venue, Concert, Ticket, DiscountCode, RecordingStudio)
+class Album(Table):
+    id: Serial
+    name = Varchar()
+    band = ForeignKey(Band)
+    release_date = Date()
+    recorded_at = ForeignKey(RecordingStudio)
+
+    @classmethod
+    def get_readable(cls) -> Readable:
+        return Readable(
+            template="%s - %s",
+            columns=[cls.name, cls.band._.name],
+        )
+
+
+TABLES = (
+    Manager,
+    Band,
+    Venue,
+    Concert,
+    Ticket,
+    DiscountCode,
+    RecordingStudio,
+    Album,
+)
 
 
 def populate():
@@ -192,24 +216,44 @@ def populate():
         *[DiscountCode({DiscountCode.code: uuid.uuid4()}) for _ in range(5)]
     ).run_sync()
 
-    RecordingStudio.insert(
-        RecordingStudio(
+    recording_studio_1 = RecordingStudio(
+        {
+            RecordingStudio.name: "Abbey Road",
+            RecordingStudio.facilities: {
+                "restaurant": True,
+                "mixing_desk": True,
+            },
+        }
+    )
+    recording_studio_1.save().run_sync()
+
+    recording_studio_2 = RecordingStudio(
+        {
+            RecordingStudio.name: "Electric Lady",
+            RecordingStudio.facilities: {
+                "restaurant": False,
+                "mixing_desk": True,
+            },
+        },
+    )
+    recording_studio_2.save().run_sync()
+
+    Album.insert(
+        Album(
             {
-                RecordingStudio.name: "Abbey Road",
-                RecordingStudio.facilities: {
-                    "restaurant": True,
-                    "mixing_desk": True,
-                },
+                Album.name: "Awesome album 1",
+                Album.recorded_at: recording_studio_1,
+                Album.band: pythonistas,
+                Album.release_date: datetime.date(year=2021, month=1, day=1),
             }
         ),
-        RecordingStudio(
+        Album(
             {
-                RecordingStudio.name: "Electric Lady",
-                RecordingStudio.facilities: {
-                    "restaurant": False,
-                    "mixing_desk": True,
-                },
-            },
+                Album.name: "Awesome album 2",
+                Album.recorded_at: recording_studio_2,
+                Album.band: rustaceans,
+                Album.release_date: datetime.date(year=2022, month=2, day=2),
+            }
         ),
     ).run_sync()