+++ piccolo/query/methods/__init__.py
@@ -9,6 +9,23 @@ from .insert import Insert
 from .objects import Objects
 from .raw import Raw
 from .refresh import Refresh
-from .select import Avg, Max, Min, Select, Sum
+from .select import Select
 from .table_exists import TableExists
 from .update import Update
+
+__all__ = (
+    "Alter",
+    "Count",
+    "Create",
+    "CreateIndex",
+    "Delete",
+    "DropIndex",
+    "Exists",
+    "Insert",
+    "Objects",
+    "Raw",
+    "Refresh",
+    "Select",
+    "TableExists",
+    "Update",
+)