+++ piccolo/querystring.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import typing as t
+from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass
 from datetime import datetime
 from importlib.util import find_spec
@@ -8,6 +9,7 @@ from string import Formatter
 
 if t.TYPE_CHECKING:  # pragma: no cover
     from piccolo.table import Table
+    from piccolo.columns import Column
 
 from uuid import UUID
 
@@ -17,22 +19,32 @@ else:
     apgUUID = UUID
 
 
-@dataclass
-class Unquoted:
+class Selectable(metaclass=ABCMeta):
     """
-    Used when we want the value to be unquoted because it's a Postgres
-    keyword - for example DEFAULT.
+    Anything which inherits from this can be used in a select query.
     """
 
-    __slots__ = ("value",)
+    __slots__ = ("_alias",)
 
-    value: str
+    _alias: t.Optional[str]
 
-    def __repr__(self):
-        return f"{self.value}"
+    @abstractmethod
+    def get_select_string(
+        self, engine_type: str, with_alias: bool = True
+    ) -> QueryString:
+        """
+        In a query, what to output after the select statement - could be a
+        column name, a sub query, a function etc. For a column it will be the
+        column name.
+        """
+        raise NotImplementedError()
 
-    def __str__(self):
-        return f"{self.value}"
+    def as_alias(self, alias: str) -> Selectable:
+        """
+        Allows column names to be changed in the result of a select.
+        """
+        self._alias = alias
+        return self
 
 
 @dataclass
@@ -42,7 +54,7 @@ class Fragment:
     no_arg: bool = False
 
 
-class QueryString:
+class QueryString(Selectable):
     """
     When we're composing complex queries, we're combining QueryStrings, rather
     than concatenating strings directly. The reason for this is QueryStrings
@@ -56,6 +68,7 @@ class QueryString:
         "query_type",
         "table",
         "_frozen_compiled_strings",
+        "columns",
     )
 
     def __init__(
@@ -64,6 +77,7 @@ class QueryString:
         *args: t.Any,
         query_type: str = "generic",
         table: t.Optional[t.Type[Table]] = None,
+        alias: t.Optional[str] = None,
     ) -> None:
         """
         :param template:
@@ -83,12 +97,42 @@ class QueryString:
 
         """
         self.template = template
-        self.args = args
         self.query_type = query_type
         self.table = table
         self._frozen_compiled_strings: t.Optional[
             t.Tuple[str, t.List[t.Any]]
         ] = None
+        self._alias = alias
+        self.args, self.columns = self.process_args(args)
+
+    def process_args(
+        self, args: t.Sequence[t.Any]
+    ) -> t.Tuple[t.Sequence[t.Any], t.Sequence[Column]]:
+        """
+        If a Column is passed in, we convert it to the name of the column
+        (including joins).
+        """
+        from piccolo.columns import Column
+
+        processed_args = []
+        columns = []
+
+        for arg in args:
+            if isinstance(arg, Column):
+                columns.append(arg)
+                arg = QueryString(
+                    f"{arg._meta.get_full_name(with_alias=False)}"
+                )
+            elif isinstance(arg, QueryString):
+                columns.extend(arg.columns)
+
+            processed_args.append(arg)
+
+        return (processed_args, columns)
+
+    def as_alias(self, alias: str) -> QueryString:
+        self._alias = alias
+        return self
 
     def __str__(self):
         """
@@ -143,7 +187,7 @@ class QueryString:
                 fragment.no_arg = True
                 bundled.append(fragment)
             else:
-                if isinstance(value, self.__class__):
+                if isinstance(value, QueryString):
                     fragment.no_arg = True
                     bundled.append(fragment)
 
@@ -195,3 +239,77 @@ class QueryString:
         self._frozen_compiled_strings = self.compile_string(
             engine_type=engine_type
         )
+
+    ###########################################################################
+
+    def get_select_string(
+        self, engine_type: str, with_alias: bool = True
+    ) -> QueryString:
+        if with_alias and self._alias:
+            return QueryString("{} AS " + self._alias, self)
+        else:
+            return self
+
+    def get_where_string(self, engine_type: str) -> QueryString:
+        return self.get_select_string(
+            engine_type=engine_type, with_alias=False
+        )
+
+    ###########################################################################
+    # Basic logic
+
+    def __eq__(self, value) -> QueryString:  # type: ignore[override]
+        return QueryString("{} = {}", self, value)
+
+    def __ne__(self, value) -> QueryString:  # type: ignore[override]
+        return QueryString("{} != {}", self, value)
+
+    def __add__(self, value) -> QueryString:
+        return QueryString("{} + {}", self, value)
+
+    def __sub__(self, value) -> QueryString:
+        return QueryString("{} - {}", self, value)
+
+    def __gt__(self, value) -> QueryString:
+        return QueryString("{} > {}", self, value)
+
+    def __ge__(self, value) -> QueryString:
+        return QueryString("{} >= {}", self, value)
+
+    def __lt__(self, value) -> QueryString:
+        return QueryString("{} < {}", self, value)
+
+    def __le__(self, value) -> QueryString:
+        return QueryString("{} <= {}", self, value)
+
+    def __truediv__(self, value) -> QueryString:
+        return QueryString("{} / {}", self, value)
+
+    def __mul__(self, value) -> QueryString:
+        return QueryString("{} * {}", self, value)
+
+    def __pow__(self, value) -> QueryString:
+        return QueryString("{} ^ {}", self, value)
+
+    def __mod__(self, value) -> QueryString:
+        return QueryString("{} % {}", self, value)
+
+    def is_in(self, value) -> QueryString:
+        return QueryString("{} IN {}", self, value)
+
+    def not_in(self, value) -> QueryString:
+        return QueryString("{} NOT IN {}", self, value)
+
+    def like(self, value: str) -> QueryString:
+        return QueryString("{} LIKE {}", self, value)
+
+    def ilike(self, value: str) -> QueryString:
+        return QueryString("{} ILIKE {}", self, value)
+
+
+class Unquoted(QueryString):
+    """
+    This is deprecated - just use QueryString directly.
+    """
+
+    pass